+++ Source/DTTableViewManager.swift
@@ -491,7 +491,7 @@ extension DTTableViewManager
     }
     
     /// Registers `closure` to be executed, when `UITableViewDelegate.tableView(_:didDeselectRowAt:)` method is called for `cellClass`.
-    open func didDeselect<T:ModelTransfer>(_ cellClass:  T.Type, _ closure: @escaping (T,T.ModelType, IndexPath) -> IndexPath?) where T:UITableViewCell {
+    open func didDeselect<T:ModelTransfer>(_ cellClass:  T.Type, _ closure: @escaping (T,T.ModelType, IndexPath) -> Void) where T:UITableViewCell {
         appendReaction(for: T.self, signature: .didDeselectRowAtIndexPath, closure: closure)
     }